add dynamic tag colors
